// gen/measurement/v1/measurement.schema.ts
export interface components {
  schemas: {
    MeasurementServiceCreateMeasurementBody: {
      measuredOn?: components["schemas"]["typeDate"];
      measuredBy?: string;
      isDraft?: boolean;
      entries?: components["schemas"]["v1MeasurementEntry"][];
    };
    MeasurementServiceUpdateMeasurementBody: {
      measuredOn?: components["schemas"]["typeDate"];
      measuredBy?: string;
      isDraft?: boolean;
      entries?: components["schemas"]["v1MeasurementEntry"][];
    };
    protobufAny: {
      "@type"?: string;
    } & {
      [key: string]: unknown;
    };
    rpcStatus: {
      /** Format: int32 */
      code?: number;
      message?: string;
      details?: components["schemas"]["protobufAny"][];
    };
    /**
     * Represents a whole or partial calendar date, such as a birthday. The time of
     *     day and time zone are either specified elsewhere or are insignificant. The
     *     date is relative to the Gregorian Calendar. This can represent one of the
     *     following:
     * @description * A full date, with non-zero year, month, and day values.
     *     * A month and day, with a zero year (for example, an anniversary).
     *     * A year on its own, with a zero month and a zero day.
     *     * A year and month, with a zero day (for example, a credit card expiration
     *       date).
     *
     *     Related types:
     *
     *     * [google.type.TimeOfDay][google.type.TimeOfDay]
     *     * [google.type.DateTime][google.type.DateTime]
     *     * [google.protobuf.Timestamp][google.protobuf.Timestamp]
     */
    typeDate: {
      /**
       * Format: int32
       * @description Year of the date. Must be from 1 to 9999, or 0 to specify a date without
       *     a year.
       */
      year?: number;
      /**
       * Format: int32
       * @description Month of a year. Must be from 1 to 12, or 0 to specify a year without a
       *     month and day.
       */
      month?: number;
      /**
       * Format: int32
       * @description Day of a month. Must be from 1 to 31 and valid for the year and month, or 0
       *     to specify a year by itself or a year and month where the day isn't
       *     significant.
       */
      day?: number;
    };
    v1CreateMeasurementResponse: {
      measurementId?: string;
    };
    v1DeleteMeasurementResponse: Record<string, never>;
    v1GetMeasurementResponse: {
      measurement?: components["schemas"]["v1Measurement"];
    };
    v1ListMeasurementsResponse: {
      measurements?: components["schemas"]["v1Measurement"][];
    };
    v1Measurement: {
      measurementId?: string;
      customerId?: string;
      measuredOn?: components["schemas"]["typeDate"];
      measuredBy?: string;
      /** Format: int64 */
      ageAtMeasurement?: number;
      updatedBy?: string;
      isDraft?: boolean;
      entries?: components["schemas"]["v1MeasurementEntry"][];
    };
    v1MeasurementEntry: {
      measurementItemId?: string;
      unmeasurable?: boolean;
      note?: string;
      values?: components["schemas"]["v1MeasurementValue"][];
    };
    v1MeasurementValue: {
      /** Format: int64 */
      trialIndex?: number;
      side?: components["schemas"]["v1Side"];
      /** Format: double */
      value?: number;
      /** Format: double */
      valueSecondary?: number;
      valueChoice?: string;
    };
    /**
     * @default SIDE_UNSPECIFIED
     * @enum {string}
     */
    v1Side: "SIDE_UNSPECIFIED" | "SIDE_NONE" | "SIDE_LEFT" | "SIDE_RIGHT";
    v1UpdateMeasurementResponse: {
      measurement?: components["schemas"]["v1Measurement"];
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}